Added MRPAS implementation
Mingos' Restrictive Precise Angle Shadowcasting is an algorithm used for
doing visibility checks for traditional roguelike games.  This commit
adds an implementation of that algorithm in gdscript and a demo project
to show it in use.

For a description of the algorithm, see

http://www.roguebasin.com/index.php?title=Restrictive_Precise_Angle_Shadowcasting

For a demo project using the algorithm see

https://github.com/matt-kimball/godot-mrpas